requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.0
google-generativeai>=0.3.0
duckduckgo-search>=4.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0 